:truck: moves feature folders into modules
:truck: add test-related files to __tests__ dir
