+++ src/frontend/jest.config.ts
@@ -158,7 +158,10 @@ const config: Config = {
 	// testLocationInResults: false,
 
 	// The glob patterns Jest uses to detect test files
-	testMatch: ["**/__tests__/**/*.[jt]s?(x)", "**/?(*.)+(spec|test|steps).[tj]s?(x)"],
+	testMatch: [
+		"**/__tests__/**/?(*.)+(spec|test|steps).[tj]s?(x)",
+		"**/?(*.)+(spec|test|steps).[tj]s?(x)",
+	],
 
 	// An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
 	// testPathIgnorePatterns: [
