+++ src/frontend/src/__tests__/features/auth/logout.feature
