+++ src/frontend/src/__tests__/features/collection/delete-collection-item.feature
