+++ src/frontend/src/__tests__/stubs/collection.stub.ts
