+++ src/frontend/src/__tests__/stubs/login.stub.ts
