+++ src/frontend/src/__tests__/stubs/metadata.stub.ts
