+++ src/frontend/src/__tests__/support/SUPPORT.md
@@ -1,0 +1,3 @@
+# SUPPORT
+
+Shareable logic that helps sets up pre-requisites for tests in `beforeEach` for example.
