+++ src/frontend/src/__tests__/support/mocks/browser.ts
