+++ src/frontend/src/layouts/DashboardLayout.tsx
@@ -2,7 +2,7 @@ import { Link } from "@tanstack/router";
 import { observer } from "mobx-react";
 import { useEffect, useState } from "react";
 
-import { AuthPresenter } from "../auth/auth.presenter";
+import { AuthPresenter } from "../modules/auth/auth.presenter";
 
 import type { MetaDataViewModel } from "../shared/types/viewmodels";
 import { useInjection } from "../ioc/useInjection";
