+++ src/frontend/src/modules/auth/auth.types.ts
