+++ src/frontend/src/routing/ProtectedRoute.tsx
@@ -3,8 +3,8 @@ import dayjs from "dayjs";
 import { observer } from "mobx-react";
 import { ReactNode, useEffect } from "react";
 
-import { AuthPresenter } from "../auth/auth.presenter";
 import { useInjection } from "../ioc/useInjection";
+import { AuthPresenter } from "../modules/auth/auth.presenter";
 import { router } from "./router";
 
 type ProtectedRouteProps = {
